+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix: Token equality: `TokenDetails` does a logical comparison with chain id and address,
   instaed of object comparison. This makes TokenDetails good for ifs and hash maps. This
   adds `TokenDetails.__eq__` and `TokenDetails.__hash__`.
+- Fix `TradeSuccess.price` is in Python `Decimal`
 
 # 0.20